+++ client/src/components/User/Search/Search.jsx
@@ -19,7 +19,7 @@ import Cards2 from "../Cards2/Cards2";
 import Paginate from "../Paginate/Paginate";
 import Loading from "../Loading/Loading";
 import NotFound from "../NotFound/NotFound";
-import styles from "./Search.module.css";
+import styles from "../../../css/Search.module.css";
 
 export default function Search() {
   const dispatch = useDispatch();
@@ -157,10 +157,10 @@ export default function Search() {
             <h4>Filter by category</h4>
             <select
               className={styles.filters}
-              // value={filterByCategory}
-              // onChange={(e) => {
-              //   handleFilterByCategory(e);
-              // }}
+            // value={filterByCategory}
+            // onChange={(e) => {
+            //   handleFilterByCategory(e);
+            // }}
             >
               <option value="All">Select Category</option>
               {categories
@@ -225,7 +225,7 @@ export default function Search() {
               onChange={(e) => {
                 handlePrice(e);
               }}
-              // value={sortingBy}
+            // value={sortingBy}
             >
               <option value="All">Select Order</option>
               <option value="asc">Ascending</option>
@@ -255,39 +255,39 @@ export default function Search() {
               {!currentProducts.length ? (
                 <Loading />
               ) : typeof currentProducts[0] === "string" ? (
-              // ) : currentProducts[0] === [] ? (
+                // ) : currentProducts[0] === [] ? (
                 <NotFound />
               ) : (
-                <div className={styles.grid}>
-                {currentProducts.map((p) => 
-                   (
-                      <Link
-                        className={styles.btnName}
-                        to={"/product/" + p.id}
-                        key={p.id}
-                      >
-                        <Cards
-                          className={styles.grid}
-                          image={p.image}
-                          name={p.name}
-                          price={p.price}
-                          offerPrice={p.offerPrice}
+                    <div className={styles.grid}>
+                      {currentProducts.map((p) =>
+                      (
+                        <Link
+                          className={styles.btnName}
+                          to={"/product/" + p.id}
+                          key={p.id}
+                        >
+                          <Cards
+                            className={styles.grid}
+                            image={p.image}
+                            name={p.name}
+                            price={p.price}
+                            offerPrice={p.offerPrice}
+                          />
+                        </Link>
+                      )
+                      )}
+                      <div className={styles.paginate}>
+                        <Paginate
+                          productsPerPage={productsPerPage}
+                          allProducts={allProducts.length}
+                          page={page}
+                          currentPage={currentPage}
+                          setCurrentPage={setCurrentPage}
                         />
-                      </Link>
-                  )
-                )}
-                <div className={styles.paginate}>
-                <Paginate
-                  productsPerPage={productsPerPage}
-                  allProducts={allProducts.length}
-                  page={page}
-                  currentPage={currentPage}
-                  setCurrentPage={setCurrentPage}
-                />
-              </div>
-                </div>
-              )}
-              
+                      </div>
+                    </div>
+                  )}
+
             </div>
             {/* <div>
               <Paginate
