+++ client/src/components/User/Second Banner/SecondBanner.jsx
@@ -1,14 +1,13 @@
 import React from "react";
 import BannerGif from "../../../assets/BannerGif.gif";
-import styles from "./SecondBanner.module.css";
-import {Link} from "react-router-dom"
-
+import styles from "../../../css/SecondBanner.module.css";
+import { Link } from "react-router-dom"
 
 export default function SecondBanner() {
     return (
         <div className={styles.main}>
-        <Link to="/shop">
-            <img src={BannerGif} alt="img" height="340px" width="100%"/>
+            <Link to="/shop">
+                <img src={BannerGif} alt="img" height="340px" width="100%" />
             </Link>
             {/* <div>
                 <button className={styles.btn}>Shop Now</button>
