--- client/src/components/User/Slider/FeaturedProducts/FeaturedProducts.module.css
@@ -1,16 +1,0 @@
-
-.grid{
-    display: grid;
-    padding: 0px 220px;
-    /* -ms-grid-column-align: center; */
-    grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: 1fr;
-    margin-left: 70px;
-    
-    /* height: -500px; */
-    /* grid-gap: 2px; */
-  }
-
-.btnName{
-    text-decoration: none;
-  }
