+++ client/src/css/SecondBanner.module.css
