+++ client/src/css/Slider.module.css
